requirements.txt
streamlit>=1.30
pytest>=7.4
python-dotenv>=1.0
pandas>=2.0
openai>=1.30.0
pypdf>=6.0.0
scikit-learn>=1.4.0